refactor: name non-storage modules by content

effort: max
